+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 Unreleased Changes
 ------------------
 
+* Fix bug matching query params that begin with `http`.
+
 v1.0.0 (2018-11-26)
 --------------------